% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dcook_gamma}
\alias{dcook_gamma}
\title{Gráfico da distância de Cook para o modelo Gamma.}
\usage{
dcook_gamma(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico da distância de Cook para o modelo Gamma.
}
\details{
A linha tracejada representa o corte proposto por Bollen, Kenneth et al de
4/n em que n é o número de observações.
}
\examples{
clotting <- data.frame(
   u = c(5,10,15,20,30,40,60,80,100),
   lot1 = c(118,58,42,35,27,25,21,19,18),
   lot2 = c(69,35,26,21,18,16,13,12,12))
modelo <- glm(lot1 ~ log(u), data = clotting, family = Gamma)
modelo \%>\% dcook_gamma()
}

