% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dcook_norm}
\alias{dcook_norm}
\title{Gráfico da distância de Cook para o modelo normal}
\usage{
dcook_norm(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico da distância de Cook para o modelo normal
}
\details{
A linha tracejada representa o corte proposto por Bollen, Kenneth et al de
4/n em que n é o número de observações.
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% dcook_norm()
}

