% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_bino}
\alias{envel_bino}
\title{Gráfico Quantil-Quantil com envelope para a distribuição Poisson.}
\usage{
envel_bino(modelo)
}
\arguments{
\item{modelo}{modelo normal ajustado}
}
\description{
Gráfico Quantil-Quantil com envelope para a distribuição Poisson.
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
modelo <- glm(counts ~ outcome + treatment, family = poisson())
modelo \%>\% envel_bino()
}

