% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_bino_}
\alias{envel_bino_}
\title{Data frame para o grafico quantil quantil com banda de confiança para o modelo
poisson}
\usage{
envel_bino_(fit.model)
}
\description{
Data frame para o grafico quantil quantil com banda de confiança para o modelo
poisson
}

