% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_gamma}
\alias{envel_gamma}
\title{Gráfico Quantil-Quantil com envelope para a distribuição Gamma.}
\usage{
envel_gamma(modelo, link = "log")
}
\arguments{
\item{modelo}{modelo normal ajustado}

\item{link}{ligacao utilizada no modelo, default é "log".}
}
\description{
Gráfico Quantil-Quantil com envelope para a distribuição Gamma.
}
\examples{
clotting <- data.frame(
   u = c(5,10,15,20,30,40,60,80,100),
   lot1 = c(118,58,42,35,27,25,21,19,18),
   lot2 = c(69,35,26,21,18,16,13,12,12))
modelo <- glm(lot1 ~ log(u), data = clotting, family = Gamma)
envel_gamma(modelo)
modelo \%>\% envel_gamma()
}

