% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_gamma_}
\alias{envel_gamma_}
\title{Data frame para o grafico quantil quantil com banda de confiança para o modelo
gama}
\usage{
envel_gamma_(fit.model)
}
\description{
Data frame para o grafico quantil quantil com banda de confiança para o modelo
gama
}

