% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_norm}
\alias{envel_norm}
\title{Gráfico Quantil-Quanntil com envelope.}
\usage{
envel_norm(modelo, dados = fit.model$model)
}
\arguments{
\item{modelo}{modelo normal ajustado utilizando o comando lm}

\item{dados}{dados utilizados para ajustar o modelo, se não for especificado
o programa utilizará modelo$model}
}
\description{
Gráfico Quantil-Quanntil com envelope.
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
envel_norm(modelo)
modelo \%>\% envel_norm()
}

