% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_pois}
\alias{envel_pois}
\title{Gráfico Quantil-Quantil com envelope para a distribuição Poisson.}
\usage{
envel_pois(modelo, link = "log")
}
\arguments{
\item{modelo}{modelo normal ajustado}

\item{link}{ligacao utilizada no modelo, default é "log".}
}
\description{
Gráfico Quantil-Quantil com envelope para a distribuição Poisson.
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
modelo <- glm(counts ~ outcome + treatment, family = poisson())
modelo \%>\% envel_pois()
}

