% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_pois_}
\alias{envel_pois_}
\title{Data frame para o grafico quantil quantil com banda de confiança para o modelo
poisson}
\usage{
envel_pois_(fit.model, link = "log")
}
\description{
Data frame para o grafico quantil quantil com banda de confiança para o modelo
poisson
}

