% Generated by roxygen2 (4.0.1): do not edit by hand
\name{idf_todas}
\alias{idf_todas}
\title{Identificar todas observações que estão fora do intervalo}
\usage{
idf_todas(p, size = 5)
}
\arguments{
\item{p}{um gráfico de diagnóstico}

\item{size}{tamanho da letra do texto, default = 5}
}
\description{
Identificar todas observações que estão fora do intervalo
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% envel_norm() \%>\% idf_todas()
}

