% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infl_bino}
\alias{infl_bino}
\title{Gráfico de influência para o modelo Poisson}
\usage{
infl_bino(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico de influência para o modelo Poisson
}
\examples{
cuse <- read.table("http://data.princeton.edu/wws509/datasets/cuse.dat",
    header=TRUE)
modelo <- glm(cbind(using, notUsing) ~age + education + wantsMore ,
                 family = binomial, data = cuse)
modelo \%>\% infl_bino()
}

