% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infl_norm}
\alias{infl_norm}
\title{Gráfico de influência para o modelo normal}
\usage{
infl_norm(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico de influência para o modelo normal
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% infl_norm()
}

