% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infl_pois}
\alias{infl_pois}
\title{Gráfico de influência para o modelo Poisson}
\usage{
infl_pois(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico de influência para o modelo Poisson
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
modelo <- glm(counts ~ outcome + treatment, family = poisson())
modelo \%>\% infl_pois()
}

