% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resind_norm}
\alias{resind_norm}
\title{Gráfico do índice pelo resíduo padronizado}
\usage{
resind_norm(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico do índice pelo resíduo padronizado
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% resind_norm()
}

