% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resva_norm}
\alias{resva_norm}
\title{Gráfico do valor ajustado pelo resíduo padronizado}
\usage{
resva_norm(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gráfico do valor ajustado pelo resíduo padronizado
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% resva_norm()
}

