% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dcook_norm}
\alias{dcook_norm}
\title{Gr치fico da dist칙ncia de Cook para o modelo normal}
\usage{
dcook_norm(modelo)
}
\arguments{
\item{modelo}{modelo ajustado}
}
\description{
Gr치fico da dist칙ncia de Cook para o modelo normal
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
modelo \%>\% dcook_norm()
}

