% Generated by roxygen2 (4.0.1): do not edit by hand
\name{envel_norm}
\alias{envel_norm}
\title{Gráfico Quantil-Quantil com envelope.}
\usage{
envel_norm(modelo)
}
\arguments{
\item{modelo}{modelo normal ajustado utilizando o comando lm}
}
\description{
Gráfico Quantil-Quantil com envelope.
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
envel_norm(modelo)
modelo \%>\% envel_norm()
}

