% Generated by roxygen2 (4.0.1): do not edit by hand
\name{idf_indice}
\alias{idf_indice}
\title{Identificar observação pelo índice}
\usage{
idf_indice(p, id)
}
\arguments{
\item{p}{um gráfico de diagnóstico de resíduos}

\item{id}{índice da observação que deseja marcar no gráfico}
}
\description{
Identificar observação pelo índice
}
\examples{
modelo <- lm(mpg ~ cyl + disp, data = mtcars)
envel_norm(modelo)
modelo \%>\% envel_norm() \%>\% idf_indice(5)
}

